requirements.txt

fastapi==0.112.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
requests==2.32.3
urllib3==2.2.2
uvicorn~=0.30.5
email_validator==2.2.0
python-dotenv==1.0.1
fastapi-pagination==0.12.26



sqlmodel~=0.0.21
SQLAlchemy~=2.0.32